+++ tasks.py
@@ -145,6 +145,7 @@ def init(context, py=None):
 def wheel(context, py=None):
     """Build a wheel."""
     print(f"Installing dependencies into: {_DEFAULT_VENV}")
+    context.run(f"{get_py(py)} setup.py sdist")
     context.run(f"{get_py(py)} setup.py bdist_wheel")